The primary change here is adding a few calls to `useMemo` through the `tableFields.tsx` file, which prevent a number of the cells from re-rendering unless their props change. The penalty of re-rendering seems to be most pronounced with the Instructor and Meeting cells, but I've added `useMemo` around all of the cells as a bit of future-proofing. 

Doing some quick testing with React Devtools add-on in Firefox:

- On `develop`, opening or closing a modal typically took around 1500ms of rendering time (+/- 100ms)
- With these changes, opening or closing takes about 450ms (+/- 50ms)

That's enough of an improvement to make the system feel less "broken", but it's still enough to feel laggy. It also only improves the render time after the initial render -- the initial page load still takes the full amount of rendering time. 

To gain further improvements, we'd need to reduce the number of rows rendered in the table. Even though blocking unchanged cells from re-rendering, it seems there is still something of a performance penalty for even checking the memoization. The tests above were done with a table containing 477 rows, so as an experiment I re-ran the profiling with these changes while also arbitrarily cutting down the number of rows:

- With 100 rows, opening or closing a modal take about 100ms (+/- 20ms)
- With 50 rows, opening or closing a modal takes about 70ms (+/- 20ms)
- With 10 rows, opening or closing a modal takes about 30ms (+/- 10ms)

From a user standpoint, I think 50 is right about the sweet spot. But as we've discussed, we'll hold off on doing any further optimization until post-MVP.

(It's also worth noting that filtering the table will also reduce the number of rows that need to be rendered/evaluated/re-rendered, so if a user is only looking at CS courses, the table will probably be a bit smoother too).

## Types of changes
<!--
  What types of changes does your code introduce? Put an `x` in all the boxes
  that apply:
-->
- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality
      to change)

## Checklist:
<!--
  Go over all the following points, and put an `x` in all the boxes that apply.
-->
- [x] I have run `eslint` on the code
- [x] I have added JSDoc for all of my code (where applicable)
- [x] I have added tests to cover my changes.

## Priority:
- [x] Normal <!-- New piece of functionality -->
- [ ] High <!-- Critical bug requiring urgent review -->

## Related Issues:
<!--
  Add the issue number in below that this PR is intended to address.

  Also mention any issues that this PR is related to, and if possible,
  provide more information regarding the relationship to the issues in
  question (i.e does the PR fix the issue, is it designed to fix another bug
  that is similar to the issue etc.)
-->
Fixes #443 

<!--
  Attribution:
  PR Template adapted from: https://github.com/h5bp/html5-boilerplate/blob/master/.github/PULL_REQUEST_TEMPLATE.md
-->